config: remove static files in etc,lib,usr,var
A lot of static files came from config/ directory which contains
a ubuntu-core-config deb package. These static files are now part
of the core snap, see:
https://github.com/snapcore/core/pull/83

So we can remove them here. This also moves "core" closer to how
core18 (and core20) are build and avoids having to upload a
package to a PPA to change e.g. the writable-path handling.